refactor: enums on different file
