+++ src/enums.ts
@@ -1,0 +1,27 @@
+/** @desc The possible sizes of a creature. */
+export enum Size {
+  Tiny = 'Tiny',
+  Small = 'Small',
+  Medium = 'Medium',
+  Large = 'Large',
+  Huge = 'Huge',
+  Gargantuan = 'Gargantuan'
+};
+
+/** @desc All the possible kinds of creatures. */
+export enum CreatureType {
+  Aberration = 'Aberration',
+  Beast = 'Beast',
+  Celestial = 'Celestial',
+  Construct = 'Construct',
+  Dragon = 'Dragon',
+  Elemental = 'Elemental',
+  Fey = 'Fey',
+  Fiend = 'Fiend',
+  Giant = 'Giant',
+  Humanoid = 'Humanoid',
+  Monstrosity = 'Monstrosity',
+  Ooze = 'Ooze',
+  Plant = 'Plant',
+  Undead = 'Undead'
+};
