+++ src/types.d.ts
@@ -119,16 +119,6 @@ export type Proficiencies = {
   [Key in keyof ProficiencyTypes]: ProficiencyTypes[Key][];
 };
 
-/** @desc The possible sizes of a creature. */
-export enum Size {
-  Tiny = 'Tiny',
-  Small = 'Small',
-  Medium = 'Medium',
-  Large = 'Large',
-  Huge = 'Huge',
-  Gargantuan = 'Gargantuan'
-};
-
 /** @desc An active or passive skill of a creature. */
 export interface Ability {
   /** A description of the ability. */
@@ -148,21 +138,3 @@ export interface Ability {
  * ```
  */
 export type Abilities = Record<string, Ability>;
-
-/** @desc All the possible kinds of creatures. */
-export enum CreatureType {
-  Aberration = 'Aberration',
-  Beast = 'Beast',
-  Celestial = 'Celestial',
-  Construct = 'Construct',
-  Dragon = 'Dragon',
-  Elemental = 'Elemental',
-  Fey = 'Fey',
-  Fiend = 'Fiend',
-  Giant = 'Giant',
-  Humanoid = 'Humanoid',
-  Monstrosity = 'Monstrosity',
-  Ooze = 'Ooze',
-  Plant = 'Plant',
-  Undead = 'Undead'
-};
